Implement a pack/unpack haloExchange step.
The MPI+OpenMP reference version packs atoms into a buffer and sends that
across to the appropriate rank. By implementing this in Chapel, I tend to see
better performance.

Prior to this commit, the haloExchange did a GET from a remote locale to the
current locale. This commit does a PUT from the current locale to the target
locale. I think this makes it easier to pack up the buffer and minimize
communication.

Removes 'inline' from several procedures. I was finding it difficult to read
the generated C code. Removing the inline didn't appear to impact performance.